#4 some sort of multiply blending works now.
Limitations:
Blending is relative to position of image to blend, which is good.
But on non overlapping areas white is used as multiply filter, which darkens the output picture

Solution: Picture must be padded with black to fit the canvas resolution, then no artifical darkening should happen anymore

Missing Specifications:
Removing and storing of filters might be troublesome aswell. Is not implemented yet.
--> Should the filter still work, when one of the corresponding images is deleted?

Troublesome might be the performance aswell, when pictures with filters on them are moved, because for each movement the filter result is different --> Debouncing?